Ush 1189 & 1198 - Images in the mobile app (#1085)

+++ apps/mobile-mzima-client/src/app/post/components/image-uploader/image-uploader.component.ts
@@ -1,6 +1,6 @@
 import { Component, forwardRef, Input } from '@angular/core';
 import { ControlValueAccessor, FormControl, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { SafeUrl } from '@angular/platform-browser';
+import { DomSanitizer, SafeUrl } from '@angular/platform-browser';
 import { Camera, CameraResultType, CameraSource, Photo } from '@capacitor/camera';
 import { Capacitor } from '@capacitor/core';
 import { Directory, FileInfo, Filesystem } from '@capacitor/filesystem';
@@ -28,6 +28,9 @@ interface LocalFile {
   ],
 })
 export class ImageUploaderComponent implements ControlValueAccessor {
+  constructor(domSanitizer: DomSanitizer) {
+    this.domSanitizer = domSanitizer;
+  }
   @Input() public hasCaption: boolean;
   @Input() public requiredError?: boolean;
   @Input() public isConnection: boolean;
@@ -41,6 +44,7 @@ export class ImageUploaderComponent implements ControlValueAccessor {
   upload = false;
   onChange: any = () => {};
   onTouched: any = () => {};
+  domSanitizer: DomSanitizer;
 
   writeValue(obj: any): void {
     if (obj) {
@@ -97,16 +101,18 @@ export class ImageUploaderComponent implements ControlValueAccessor {
     this.fileName = new Date().getTime() + '.jpeg';
     const filePath = `${IMAGE_DIR}/${this.fileName}`;
     try {
-      const savedFile = await Filesystem.writeFile({
-        directory: Directory.Data,
-        path: filePath,
-        data: base64Data,
-      });
       // const file = await this.loadFile();
 
-      if (Capacitor.getPlatform() === 'hybrid') {
+      if (['hybrid'].includes(Capacitor.getPlatform())) {
         // Display the new image by rewriting the 'file://' path to HTTP
         // Details: https://ionicframework.com/docs/building/webview#file-protocol
+
+        const savedFile = await Filesystem.writeFile({
+          directory: Directory.Data,
+          path: filePath,
+          data: base64Data,
+        });
+
         this.photo = {
           name: this.fileName,
           path: savedFile.uri,
@@ -124,7 +130,7 @@ export class ImageUploaderComponent implements ControlValueAccessor {
       }
 
       this.upload = true;
-      this.preview = null;
+      // this.preview = photo;
     } catch (e) {
       console.log(e);
     }
@@ -132,10 +138,12 @@ export class ImageUploaderComponent implements ControlValueAccessor {
 
   async deleteSelectedImage() {
     try {
-      await Filesystem.deleteFile({
-        directory: Directory.Data,
-        path: this.photo!.path,
-      });
+      if (Capacitor.getPlatform() === 'hybrid') {
+        await Filesystem.deleteFile({
+          directory: Directory.Data,
+          path: this.photo!.path,
+        });
+      }
       this.photo = null;
       this.upload = false;
       this.preview = null;
